Use bindgen to generate sys crate (0.69.0)
While hand-generated bindings are nice, the size of the Flipper Zero
SDK (~1,800 public symbols) isn't something that can be feasibly kept up
to date.

Thankfully `flipperzero-firmware` has introduced SDK support, which
makes it substantially easier to ensure that we have the right compile
flags and headers included.

While it's possible to build the sys-bindings using a build script,
I've instead opted to check the bindings in and provide a
`generate-bindings` program for updates. This means that users don't
need to have a copy of `flipperzero-firmware` with a supported toolchain
to use the crate.

Due to the way Cargo targets work, I've had to move all
`thumbv7em-none-eabihf` code into a separate sub-directory as it does
not appear possible to override build targets when using workspaces
(either to unset a previously defined target, or to override a target
in a workspace subproject).