rt: Don't wait for threads ending in "Srv"
When checking to see if there are any lingering threads, we use
a thread's `appid` to determine if it belongs to this app.

Unfortunately `appid` is not guaranteed to be unique on the system, so
it is possible for us to match one of the system apps/services.

In our case, the "gui" and "notification" examples match the name of the
built-in services, thus they would hang at exit waiting for the service
thread to terminate.

A simple workaround is to check if a thread's name ends in "Srv" (e.g. "GuiSrv)
and if so, we ignore it for the purpose of waiting for lingering threads.