flipper: Make FuriString::as_mut_ptr public
This is needed as there are certain APIs that are not accessible
through the Rust bindings (e.g. `furi_string_printf`).

This function is safe because just holding a pointer is OK as you can't
use it to perform any unsafe operations without making an unsafe action
such as dereferencing or passing to a C function.