Fix modules affected by enum representation changes
Most of the time this consists of just removing the prefix,
but sometimes it is required to explicitly convert to/from the numeric
representation (the newtype inner value).

Right at the moment none of the enums are marked as bitmask style values
which makes them a little arkward to use, but that can be improved in
a future PR. The goal here is to have the minimal changes for easy review.